+++ src/app/pages/angular/angular-routing.module.ts
@@ -9,6 +9,7 @@ const routes: Routes = [
     component: AngularComponent,
     children: [
       { path: 'life-cycle', loadChildren: () => import('./life-cycle/life-cycle.module').then((m) => m.LifeCycleModule) },
+      { path: 'ng-common', loadChildren: () => import('./ng-common/ng-common.module').then((m) => m.NgCommonModule) },
       { path: 'ng-core', loadChildren: () => import('./ng-core/ng-core.module').then((m) => m.NgCoreModule) },
       { path: 'ng-animations', loadChildren: () => import('./ng-animations/ng-animations.module').then((m) => m.NgAnimationsModule) },