+++ src/app/pages/angular/ng-common/ng-common.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'da-ng-common',
+  template: ` <router-outlet></router-outlet> `,
+})
+export class NgCommonComponent {}