+++ src/app/pages/angular/ng-common/ng-component-outlet-demo/comp-a.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-comp-a',
+  template: `
+    <h3>component A works!</h3>
+  `
+})
+export class CompAComponent {
+
+}