+++ src/app/pages/angular/ng-common/ng-component-outlet-demo/comp-b.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-comp-b',
+  template: `
+    <h3>component B works!</h3>
+  `
+})
+export class CompBComponent {
+
+}