+++ src/app/pages/angular/ng-common/ng-component-outlet-demo/comp-c.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-comp-c',
+  template: `
+    <h3>component C works!</h3>
+  `
+})
+export class CompCComponent {
+
+}