+++ src/app/pages/angular/ng-common/ng-component-outlet-demo/ng-component-outlet-demo-routing.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { NgComponentOutletDemoComponent } from './ng-component-outlet-demo.component';
+
+const routes: Routes = [
+  { path: '', component: NgComponentOutletDemoComponent }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NgComponentOutletDemoRoutingModule { }