+++ ChangeLog.md
@@ -5,4 +5,7 @@
 
 ### 2024-05-14
 1. 添加 插件 `ngx-markdown`
-2. 完成迁移 angular/生命周期 内容
+2. 完成迁移 angular - 生命周期 内容
+
+### 2024-05-14
+1. 完成迁移 angular - @angular/common 内容